Username validation in router. Validation refactor
