+++ waspc/data/Generator/templates/server/src/auth/providers/local/signup.ts
@@ -1,10 +1,12 @@
 {{={= =}=}}
 import { handleRejection } from '../../../utils.js'
 import { createUser } from '../../utils.js'
+import { ensureValidUsernameAndPassword } from '../../validation.js'
 import { validateAndGetAdditionalFields } from '../../utils.js'
 
 export default handleRejection(async (req, res) => {
   const userFields = req.body || {}
+  ensureValidUsernameAndPassword(userFields)
 
   const additionalFields = await validateAndGetAdditionalFields(userFields)
